--- app/serializers/answer_serializer.rb
@@ -1,8 +1,0 @@
-class AnswerSerializer < ActiveModel::Serializer
-  attributes :id, :body, :best_answer, :user_id
-
-  include SerializedFilable
-  include SerializedTimestampable
-  include SerializedLinkable
-  include SerializedCommentable
-end
