--- app/serializers/answers_serializer.rb
@@ -1,10 +1,0 @@
-
-class AnswersSerializer < ActiveModel::Serializer
-  attributes :id, :body, :best_answer, :user_id, :short_body
-
-  def short_body
-    object.body.truncate(10)
-  end
-
-  include SerializedTimestampable
-end
