--- app/serializers/comments_serializer.rb
@@ -1,6 +1,0 @@
-
-class CommentsSerializer < ActiveModel::Serializer
-  attributes :id, :body
-
-  include SerializedTimestampable
-end
