--- app/serializers/concerns/serialized_commentable.rb
@@ -1,7 +1,0 @@
-module SerializedCommentable
-  extend ActiveSupport::Concern
-
-  included do
-    has_many :comments, serializer: CommentsSerializer
-  end
-end
