+++ app/serializers/concerns/v1/serialized_commentable.rb
@@ -1,0 +1,9 @@
+module V1
+  module SerializedCommentable
+    extend ActiveSupport::Concern
+
+    included do
+      has_many :comments, serializer: V1::CommentsSerializer
+    end
+  end
+end
