+++ app/serializers/concerns/v1/serialized_linkable.rb
@@ -1,0 +1,10 @@
+module V1
+  module SerializedLinkable
+    extend ActiveSupport::Concern
+
+    included do
+      has_many :comments, serializer: V1::CommentsSerializer
+      has_many :links, serializer: V1::LinksSerializer
+    end
+  end
+end
