+++ app/serializers/concerns/v1/serialized_timestampable.rb
@@ -1,0 +1,18 @@
+module V1
+  module SerializedTimestampable
+    extend ActiveSupport::Concern
+
+    included do
+      attribute :created_at
+      attribute :updated_at
+    end
+
+    def created_at
+      object.created_at.as_json
+    end
+
+    def updated_at
+      object.created_at.as_json
+    end
+  end
+end
