--- app/serializers/links_serializer.rb
@@ -1,6 +1,0 @@
-
-class LinksSerializer < ActiveModel::Serializer
-  attributes :id, :name, :url
-
-  include SerializedTimestampable
-end
