--- app/serializers/question_serializer.rb
@@ -1,8 +1,0 @@
-class QuestionSerializer < ActiveModel::Serializer
-  attributes :id, :title, :body, :user_id
-
-  include SerializedFilable
-  include SerializedTimestampable
-  include SerializedLinkable
-  include SerializedCommentable
-end
