--- app/serializers/user_serializer.rb
@@ -1,6 +1,0 @@
-
-class UserSerializer < ActiveModel::Serializer
-  attributes :id, :email, :admin
-
-  include SerializedTimestampable
-end
