+++ app/serializers/v1/comments_serializer.rb
@@ -1,0 +1,7 @@
+module V1
+  class CommentsSerializer < ActiveModel::Serializer
+    attributes :id, :body
+
+    include V1::SerializedTimestampable
+  end
+end
