+++ app/serializers/v1/links_serializer.rb
@@ -1,0 +1,7 @@
+module V1
+  class LinksSerializer < ActiveModel::Serializer
+    attributes :id, :name, :url
+
+    include V1::SerializedTimestampable
+  end
+end
