+++ app/serializers/v1/user_serializer.rb
@@ -1,0 +1,7 @@
+module V1
+  class UserSerializer < ActiveModel::Serializer
+    attributes :id, :email, :admin
+
+    include V1::SerializedTimestampable
+  end
+end
