+++ spec/api/v1/profiles_spec.rb
@@ -45,7 +45,7 @@ describe 'Profiles APLI', type: :request do
       it 'returns all users' do
         expect(json['users'].size).to eq(3)
         expect(json['users']).
-          to match_array(another_records.map { |record| UserSerializer.new(record).as_json.deep_stringify_keys })
+          to match_array(another_records.map { |record| V1::UserSerializer.new(record).as_json.deep_stringify_keys })
       end
     end
   end
